fix: startConceptCatchAnimationをクラス名の付与の処理とfromToのScrollTriggerのtriggerを別にし、2つのtimelineを生成するよう修正
