+++ src/components/Concept/index.astro
@@ -59,7 +59,9 @@ import BoldText from "@/components/BoldText.astro";
 </section>
 
 <script>
-  import startConceptCatchAnimation from "./startConceptCatchAnimation";
+  import startConceptCatchFillAnimation from "./startConceptCatchFillAnimation";
+  import startConceptCatchSkewAnimation from "./startConceptCatchSkewAnimation";
 
-  startConceptCatchAnimation();
+  startConceptCatchFillAnimation();
+  startConceptCatchSkewAnimation();
 </script>
