+++ src/components/Concept/startConceptCatchSkewAnimation.ts
@@ -1,0 +1,50 @@
+import { gsap } from "gsap";
+import { ScrollTrigger } from "gsap/ScrollTrigger";
+import resetScrollTriggerTls from "@/animations/resetScrollTriggers/managers/resetScrollTriggerTls";
+
+const startConceptCatchSkewAnimation = () => {
+  const trigger = document.getElementById("concept");
+  const scrollCatchEl = document
+    .getElementById("catch-scroll")
+    ?.querySelector("span");
+
+  if (trigger && scrollCatchEl) {
+    gsap.registerPlugin(ScrollTrigger);
+
+    let isFirst: boolean = true;
+
+    const tl = gsap.timeline({
+      scrollTrigger: {
+        trigger,
+        id: "catchSkew",
+        start: "top 50%",
+        end: "bottom 80%",
+        scrub: true,
+      },
+    });
+
+    tl.fromTo(
+      scrollCatchEl,
+      {
+        skewY: 0,
+        scaleX: 1.0,
+        scaleY: 1.0,
+        rotate: 0,
+      },
+      {
+        skewY: 12,
+        scaleX: 1.05,
+        scaleY: 1.6,
+        rotate: -12,
+        onComplete: () => {
+          if (isFirst) {
+            isFirst = false;
+            resetScrollTriggerTls();
+          }
+        },
+      },
+    );
+  }
+};
+
+export default startConceptCatchSkewAnimation;
